refactor: keep npm dependency provider isolated
