+++ src/extensions/mod.rs
@@ -1,1 +1,2 @@
 pub mod deps_provider;
+pub mod npm_deps_provider;
